+++ src/mechanics/integrands/TimeDependent.h
@@ -7,7 +7,7 @@
 
 namespace NuTo
 {
-namespace Integrand
+namespace Integrands
 {
 namespace TimeDependent
 {
